bug fix
문제 풀이에는 지장없었지만, 그래도 선언한 타입과 일치하는지 항상 확인하자.

+++ www.dovelet.com/stair_03/divisor1/main.cpp
@@ -11,8 +11,8 @@ int main(int argc, char* argv[]) {
     long long A, B, a, b, sum, count, i, mod = 1000000003;
 
     while (cin >> A >> B) {
-        a = (int)sqrtl((long double)(A - 1.0)) + 1;
-        b = (int)sqrtl((long double)(B));
+        a = (long long)sqrtl((long double)(A - 1.0)) + 1;
+        b = (long long)sqrtl((long double)(B));
         sum = count = 0;
         for (i = a; i <= b; i++) {
             sum = (sum + (i % mod) * (i % mod) % mod) % mod;
